+++ react/ProductList.tsx
@@ -1,33 +1,51 @@
 import React from 'react'
 import { jsonLdScriptProps } from 'react-schemaorg'
 import { ItemList, WithContext, ListItem } from 'schema-dts'
+import { useRuntime } from 'vtex.render-runtime'
 
-import { getBaseUrl } from './modules/baseUrl'
+import useAppSettings from './hooks/useAppSettings'
+import { parseToJsonLD } from './Product'
 
 interface Product {
   productName: string
   linkText: string
   link: string
+  items: any[]
 }
 
 interface Props {
   products?: Product[]
 }
 
-export function getProductList(products?: Product[]) {
+export function getProductList({
+  decimals,
+  pricesWithTax,
+  currency,
+  products,
+}: {
+  decimals: number
+  pricesWithTax: boolean
+  currency: string
+  products?: Product[]
+}) {
   if (!Array.isArray(products)) {
     return null
   }
 
-  const baseUrl = getBaseUrl()
-
   const productItems: ListItem[] = products.map((product, index) => {
-    const productUrl = `${baseUrl}/${product.link ?? `${product.linkText}/p`}`
+    const [selectedItem] = product.items
+    const item = parseToJsonLD({
+      product,
+      selectedItem,
+      currency,
+      decimals,
+      pricesWithTax,
+    })
+
     return {
       '@type': 'ListItem',
       position: index + 1,
-      name: product.productName,
-      url: productUrl,
+      ...(item && { item }),
     }
   })
 
@@ -39,7 +57,17 @@ export function getProductList(products?: Product[]) {
 }
 
 function ProductList({ products }: Props) {
-  const productListLD: WithContext<ItemList> | null = getProductList(products)
+  const {
+    culture: { currency },
+  } = useRuntime()
+
+  const { decimals, pricesWithTax } = useAppSettings()
+  const productListLD: WithContext<ItemList> | null = getProductList({
+    decimals,
+    pricesWithTax,
+    currency,
+    products,
+  })
 
   if (!productListLD) {
     return null